Initial commit (mid-air fix and new functionality)
This initial commit fixes a painful bug with the original code. Specifically, there wasn't a check to see if the tile found via GetElevatorExit was solid. Depending on the placement of the elevator, this could result in the character being moved to a floorless tile adjacent to the elevator many floors up. I should know - this happened to me.

Elevators are now powered and operated differently compared to the original mod. Any elevator context check will now look at every floor along the route. If any of them are powered, then every floor will be interactable. Additionally, with a high enough Metalworking and Electrical skill (10 and 7, respectively) it's possible to craft new elevators.

I have future considerations now that this basic code works (realistic control conduit and motor requirements, working elevator cab with embark/disembark via vehicle code), but it's possible that those might be a separate mod.